Gosia secure FINANSE endpoints

+++ src/main/java/com/sggw/accessmux/web/rest/FINANCESResource.java
@@ -1,6 +1,7 @@
 package com.sggw.accessmux.web.rest;
 
 import com.codahale.metrics.annotation.Timed;
+import com.sggw.accessmux.security.SecurityUtils;
 import com.sggw.accessmux.service.FINANCESService;
 import com.sggw.accessmux.web.rest.errors.BadRequestAlertException;
 import com.sggw.accessmux.web.rest.util.HeaderUtil;
@@ -17,6 +18,9 @@ import java.net.URISyntaxException;
 
 import java.util.List;
 import java.util.Optional;
+
+import static com.sggw.accessmux.security.AuthoritiesConstants.ADMIN;
+import static com.sggw.accessmux.security.AuthoritiesConstants.FINANSE;
 
 /**
  * REST controller for managing FINANCES.
@@ -45,6 +49,11 @@ public class FINANCESResource {
     @PostMapping("/finances")
     @Timed
     public ResponseEntity<FINANCESDTO> createFINANCES(@Valid @RequestBody FINANCESDTO fINANCESDTO) throws URISyntaxException {
+
+        if (!(SecurityUtils.isCurrentUserInRole(FINANSE) || SecurityUtils.isCurrentUserInRole(ADMIN))) {
+            return ResponseEntity.badRequest().headers(HeaderUtil.createFailureAlert(ENTITY_NAME, "http.403","You need to be logged!")).body(null);
+        }
+
         log.debug("REST request to save FINANCES : {}", fINANCESDTO);
         if (fINANCESDTO.getId() != null) {
             throw new BadRequestAlertException("A new fINANCES cannot already have an ID", ENTITY_NAME, "idexists");
@@ -67,6 +76,11 @@ public class FINANCESResource {
     @PutMapping("/finances")
     @Timed
     public ResponseEntity<FINANCESDTO> updateFINANCES(@Valid @RequestBody FINANCESDTO fINANCESDTO) throws URISyntaxException {
+
+        if (!(SecurityUtils.isCurrentUserInRole(FINANSE) || SecurityUtils.isCurrentUserInRole(ADMIN))) {
+            return ResponseEntity.badRequest().headers(HeaderUtil.createFailureAlert(ENTITY_NAME, "http.403","You need to be logged!")).body(null);
+        }
+
         log.debug("REST request to update FINANCES : {}", fINANCESDTO);
         if (fINANCESDTO.getId() == null) {
             throw new BadRequestAlertException("Invalid id", ENTITY_NAME, "idnull");
@@ -85,6 +99,11 @@ public class FINANCESResource {
     @GetMapping("/finances")
     @Timed
     public List<FINANCESDTO> getAllFINANCES() {
+
+        if (!(SecurityUtils.isCurrentUserInRole(FINANSE) || SecurityUtils.isCurrentUserInRole(ADMIN))) {
+            return null;
+        }
+
         log.debug("REST request to get all FINANCES");
         return fINANCESService.findAll();
     }
@@ -98,6 +117,11 @@ public class FINANCESResource {
     @GetMapping("/finances/{id}")
     @Timed
     public ResponseEntity<FINANCESDTO> getFINANCES(@PathVariable Long id) {
+
+        if (!(SecurityUtils.isCurrentUserInRole(FINANSE) || SecurityUtils.isCurrentUserInRole(ADMIN))) {
+            return ResponseEntity.badRequest().headers(HeaderUtil.createFailureAlert(ENTITY_NAME, "http.403","You need to be logged!")).body(null);
+        }
+
         log.debug("REST request to get FINANCES : {}", id);
         Optional<FINANCESDTO> fINANCESDTO = fINANCESService.findOne(id);
         return ResponseUtil.wrapOrNotFound(fINANCESDTO);
@@ -112,6 +136,11 @@ public class FINANCESResource {
     @DeleteMapping("/finances/{id}")
     @Timed
     public ResponseEntity<Void> deleteFINANCES(@PathVariable Long id) {
+
+        if (!(SecurityUtils.isCurrentUserInRole(FINANSE) || SecurityUtils.isCurrentUserInRole(ADMIN))) {
+            return ResponseEntity.badRequest().headers(HeaderUtil.createFailureAlert(ENTITY_NAME, "http.403","You need to be logged!")).body(null);
+        }
+
         log.debug("REST request to delete FINANCES : {}", id);
         fINANCESService.delete(id);
         return ResponseEntity.ok().headers(HeaderUtil.createEntityDeletionAlert(ENTITY_NAME, id.toString())).build();
